feat: Add endpoint for retrieving messages
This commit adds a new endpoint `/messages` to the backend server. When accessed, it returns a JSON response containing an array of messages. This endpoint is intended to be used by the frontend to fetch messages to display to the user.

The purpose of this change is to provide the necessary backend functionality for the frontend to retrieve and display messages.